+++ packages/pl-fe/src/stores/settings.ts
@@ -65,7 +65,6 @@ const useSettingsStore = create<State>()(mutative((set) => ({
     changeSetting(state.userSettings, path, value);
 
     mergeSettings(state);
-    console.log(JSON.parse(JSON.stringify(state.userSettings)));
   }),
 
   rememberEmojiUse: (emoji: Emoji) => set((state: State) => {